modularize form business and owner information
