+++ src/index.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import FormContainer from './components/form_container';
 import './index.css';
+
+import LoanApp from './loanApp';
 
 ReactDOM.render(
     <div>
-        <FormContainer />
+        <LoanApp />
     </div>,
     document.getElementById('root')
 );
